backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt<4.0.0
PyJWT
email-validator
python-multipart
